refactor: move ALL of result_utils to search.py
